Tidy/update/move examples and tests.
Examples are tidied up and moved to prism-examples/smgs.
Where files have been deleted, they should be available in the
web page accompanying the relevant paper (see each README for the link).
Compositional examples are retained in the compo branch.

Tests are now stable and moved from prism/tests to prism-tests.
More specifically, they are in:

* prism-tests/functionality/verify/smgs
* prism-tests/functionality/verify/smgs-multi

where the latter set use multi-objective techniques and require PPL to be
set up and installed correctly to work.