Simplify dependencies on PPL.
During compilation, the required libraries are copied into the local
lib directory, making it self-contained at run time.
This means that PPL_DIR does not need to be set in the launch scripts,
either when compiling from source, or installing a binary release.
Similarly, the sub-Makefiles do not need PPL_DIR.