Array sup-norm methods in PrismUtils consistently check only over common elements.
Assertions that checked that, e.g., arrays were the same size have been removed.
(These are disabled in the JDK by default anyway)

Some value iteration methods, which take in an "init" array of known values
can violate these assertions, e.g., when it is created to be over size for efficiency
in an abstraction refinement loop.